+++ clib/src/mmtrace.c
@@ -1,4 +1,4 @@
-/* @(#)root/clib:$Name$:$Id$ */
+/* @(#)root/clib:$Name:  $:$Id: mmtrace.c,v 1.1.1.1 2000/05/16 17:00:43 rdm Exp $ */
 /* Author: */
 
 /* More debugging hooks for `mmalloc'.
@@ -89,7 +89,7 @@ tr_mallochook (md, size)
   mdp -> mmalloc_hook = tr_mallochook;
 
   /* We could be printing a NULL here; that's OK.  */
-  fprintf (mallstream, "+ %08lx %x\n", (unsigned long) hdr, size);
+  fprintf (mallstream, "+ %08lx %x\n", (unsigned long) hdr, (unsigned) size);
 
   if (hdr == mallwatch)
     tr_break ();
@@ -120,10 +120,10 @@ tr_reallochook (md, ptr, size)
   mdp -> mrealloc_hook = tr_reallochook;
   if (hdr == NULL)
     /* Failed realloc.  */
-    fprintf (mallstream, "! %08lx %x\n", (unsigned long) ptr, size);
+    fprintf (mallstream, "! %08lx %x\n", (unsigned long) ptr, (unsigned) size);
   else
     fprintf (mallstream, "< %08lx\n> %08lx %x\n", (unsigned long) ptr,
-	     (unsigned long) hdr, size);
+	     (unsigned long) hdr, (unsigned) size);
 
   if (hdr == mallwatch)
     tr_break ();
